chore: move user and assistance request management
