+++ web-client/src/pages/Admin/TeamManagement/components/TeamItem.tsx
@@ -1,6 +1,6 @@
 import { memo } from "react";
 import { useParams, useNavigate } from "react-router-dom";
-import { TTeam } from "../Types/Team";
+import { TTeam } from "../../../../types/team";
 import { PiWarningCircleFill } from "react-icons/pi";
 type TProps = {
   team: TTeam;
@@ -14,9 +14,8 @@ const TeamItem: React.FC<TProps> = memo(({ team }) => {
 
   return (
     <div
-      className={`my-2 p-2 rounded-md bg-gray-100 cursor-pointer hover:bg-gray-200 transition-all duration-100 ${
-        id ? (id === team._id ? "bg-gray-200" : "") : ""
-      }`}
+      className={`my-2 p-2 rounded-md bg-gray-100 cursor-pointer hover:bg-gray-200 transition-all duration-100 ${id ? (id === team._id ? "bg-gray-200" : "") : ""
+        }`}
       key={team._id}
       onClick={() => {
         navigate(`${team._id}`);
