+++ web-client/src/types/team.d.ts
@@ -1,4 +1,4 @@
-import { User } from "../../../types/user";
+import { User } from "./user";
 
 export type TTeam = {
   _id: string;
